feat(data table manager): split data table stories
