+++ packages/components/data-table-manager/src/stories/items-data.ts
@@ -1,0 +1,122 @@
+export const items = [
+  {
+    id: '5e188c29791747d9c54250e2',
+    isExpanded: false,
+    name: 'Morgan Bean',
+    customRenderer: 'CYCLONICA',
+    phone: '+1 (895) 529-3300',
+    age: 23,
+    about:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+    itemDetails:
+      'More details rendered in the nested item. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+  },
+  {
+    id: '5e188c295ae0bb19afbb115f',
+    isExpanded: false,
+    name: 'Franklin Cochran',
+    customRenderer: 'TINGLES',
+    phone: '+1 (835) 571-3268',
+    age: 36,
+    about:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+    itemDetails:
+      'More details rendered in the nested item. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+  },
+  {
+    id: '5e188c298f0ea901553c517f',
+    isExpanded: false,
+    name: 'Salazar Craig',
+    customRenderer: 'ECRAZE',
+    phone: '+1 (944) 445-2594',
+    age: 21,
+    about:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+    itemDetails:
+      'More details rendered in the nested item. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+  },
+  {
+    id: '5e188c29b09bb748df833ed0',
+    isExpanded: false,
+    name: 'Pamela Noble',
+    customRenderer: 'FILODYNE',
+    phone: '+1 (875) 421-3328',
+    age: 34,
+    about:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+    itemDetails:
+      'More details rendered in the nested item. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+  },
+  {
+    id: '5e188c29bc14e3b97ab2ad7d',
+    isExpanded: false,
+    name: 'Terra Morrow',
+    customRenderer: 'DAISU',
+    phone: '+1 (807) 436-2026',
+    age: 30,
+    about:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+    itemDetails:
+      'More details rendered in the nested item. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+  },
+  {
+    id: '5e188c296c9b7cf486a0479c',
+    isExpanded: false,
+    name: 'Cline Hansen',
+    customRenderer: 'ULTRIMAX',
+    phone: '+1 (934) 402-3675',
+    age: 21,
+    about:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+    itemDetails:
+      'More details rendered in the nested item. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+  },
+  {
+    id: '5e188c29b45c669d8e60303f',
+    isExpanded: false,
+    name: 'Jefferson Rosario',
+    customRenderer: 'COMTOURS',
+    phone: '+1 (874) 437-2581',
+    age: 32,
+    about:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+    itemDetails:
+      'More details rendered in the nested item. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+  },
+  {
+    id: '5e188c29ca865647af147b4a',
+    isExpanded: false,
+    name: 'Tania Waller',
+    customRenderer: 'DOGSPA',
+    phone: '+1 (964) 585-3040',
+    age: 35,
+    about:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+    itemDetails:
+      'More details rendered in the nested item. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+  },
+  {
+    id: '5e188c2910b83f907e9c66ab',
+    isExpanded: false,
+    name: 'Butler Shepard',
+    customRenderer: 'HOUSEDOWN',
+    phone: '+1 (888) 434-2153',
+    age: 21,
+    about:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+    itemDetails:
+      'More details rendered in the nested item. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+  },
+  {
+    id: '5e188c29a9ece9123d6a87a1',
+    isExpanded: false,
+    name: 'Diana Wise',
+    customRenderer: 'SPEEDBOLT',
+    phone: '+1 (992) 535-2912',
+    age: 27,
+    about:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+    itemDetails:
+      'More details rendered in the nested item. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Arcu dictum varius duis at consectetur lorem donec.',
+  },
+];
